requirements.txt
discord.py
proxmoxer
urllib3
requests
matplotlib
